Add files via upload
Thank you for your feedback. I have made the change to the Starting Price to get the Returns for 2017 and 2018 that are required. I had used the starting price as the price of the ticker on the opening day (column 3) but the instruction material uses the ending price for that stock ticker on day 1 (col 6) as the starting price. I may not have followed this properly. These Returns now match the numbers shown in the instruction materials.